preprocessing/pyproject.toml
[project]
name = "agriguard-preprocessing"
version = "0.1.0"
description = "Data preprocessing for AgriGuard crop stress detection"
authors = [
    {name = "AgriGuard Team", email = "agriguard@example.com"}
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "pyarrow>=14.0.0",
    "google-cloud-storage>=2.10.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "ipython>=8.14.0",
    "jupyter>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ADD THIS SECTION ⬇️
[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]